latency

+++ core/include/coco/util/timing.h
@@ -54,7 +54,7 @@ namespace coco
 namespace util
 {
 
-inline double time()
+inline int long time()
 {
     return std::chrono::duration_cast<std::chrono::microseconds>(
             std::chrono::system_clock::now().time_since_epoch()).count();
